* master: (81 commits)
  pom: remove spam
  pom: generate exclusions
  push: disallow empty project name & revision
  require non-empty project name and revision
  Default database connection pool size should be larger on a large
machine
  start 0.8.15
  release 0.8.14
  td operator: handle empty query result
  added comment on ServerBootstrap.startServer
  scheduler command shouldn't override database.type set by -c config
file [#132]
  session status path should be at project path, not current directory
  fix --task-log usage message
  update error message formatting style
  Adds logout url config option
  Adds 404 handler page
  Revert the previous change in README.md
  updated digdag-plugin-example
  fix dependency for maven test-compile
  fix dependency for maven test-compile
  create docker tmp dir
  ...

# Conflicts:
#	digdag-ui/console.jsx
#	digdag-ui/public/config.js

+++ digdag-core/src/main/java/io/digdag/core/schedule/ScheduleControlStore.java
@@ -5,7 +5,7 @@ import io.digdag.spi.ScheduleTime;
 
 public interface ScheduleControlStore
 {
-    boolean updateNextScheduleTime(long schedId, ScheduleTime nextTime, Instant lastSessionTime);
+    boolean updateNextScheduleTime(int schedId, ScheduleTime nextTime, Instant lastSessionTime);
 
-    boolean updateNextScheduleTime(long schedId, ScheduleTime nextTime);
+    boolean updateNextScheduleTime(int schedId, ScheduleTime nextTime);
 }
